+++ ui/src/components/SelectImportPage/index.tsx
@@ -18,7 +18,8 @@ import { ConfirmationScreen } from './screens/ConfirmationScreen';
 import { SelectorItem } from './screens/SelectProjectsScreen/buildGroupsSelectorOptions';
 import { useAppContext } from '../../hooks/useAppContext';
 import { useComponentTypes } from '../../hooks/useComponentTypes';
-import { checkOnboardingRedirection, getComponentTypeOption } from '../utils';
+import { getComponentTypeOption } from '../utils';
+import { checkOnboardingRedirection } from '../onboarding-flow-context-helper';
 import { getAvailableImportComponentTypes } from './utils';
 import { useProjects } from '../../hooks/useProjects';
 import { useTeamsForImport } from '../../hooks/useTeamsForImport';
